delete_last_checkpoint.f90: small bugfix

+++ src/delete_last_checkpoint.f90
@@ -2,11 +2,19 @@ subroutine delete_last_checkpoint
 use variables
 implicit none
 character(100) :: temporary_filename
+logical :: checkpoint_exists
 
+! create end_of_simulation.csv to skip this simulation when restarting multiple parallel simulations that had been -
+! simulations restart from the checkpoint of some job of larger index that has not finished
 write(temporary_filename, '(A, "/end_of_simulation.csv")') trim(output_directory)
 open(unit=90, file=temporary_filename)
 write(90, '(A)') 'simulation is complete'
 close(90)
-call system('rm ' // trim(checkpoint_filename))
+
+write(checkpoint_filename, '(A, "/checkpoint.csv")') trim(output_directory)
+inquire(file=checkpoint_filename, exist=checkpoint_exists)
+if (checkpoint_exists) then
+    call system('rm ' // trim(checkpoint_filename))
+end if
 
 end subroutine delete_last_checkpoint
